Move PaymentMethodViewModel out of DirectCheckoutViewModel. Make surcharge a coomputed var.
